refactor: Improve Quality Gate
* Duplicate CompanyConverter code
* Lambda to method reference
* use .toList() references
* Throw allert
