+++ src/test/java/br/com/b3/service/datacharge/CompanyConverterTest.java
@@ -1,5 +1,8 @@
 package br.com.b3.service.datacharge;
 
+import br.com.b3.entity.Company;
+import br.com.b3.service.converter.CompanyConverter;
+import br.com.b3.service.dto.CompanyResponse;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -7,8 +10,9 @@ class CompanyConverterTest {
 
     @Test
     public void newInstance(){
-        CompanyConverter company = new CompanyConverter();
+        CompanyResponse response = new CompanyResponse(1L, "test", "TST11", 100.00);
+        Company company = CompanyConverter.convert(response);
+
         Assertions.assertThat(company).isNotNull();
-        Assertions.assertThat(company).isInstanceOf(CompanyConverter.class);
     }
 }
